+++ tests/fixtures/dd/gnudd-conv-sync-ibs-521-obs-1031-deadbeef.spec
