+++ tests/fixtures/dd/lcase-ascii.test
