+++ tests/fixtures/dd/lcase-ebcdic.spec
