+++ tests/fixtures/dd/lcase-ebcdic.test
