+++ tests/fixtures/dd/lcase-ibm.spec
