+++ tests/fixtures/dd/lcase-ibm.test
