+++ tests/fixtures/dd/seq-byte-values-odd.spec
