+++ tests/fixtures/dd/ucase-ascii.test
