+++ tests/fixtures/dd/ucase-ebcdic.spec
