+++ tests/fixtures/dd/ucase-ibm.spec
